push pc Miguel

+++ includes/cub3d_struct.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   cub3d_struct.h                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mescobar <mescobar@student.42.fr>          +#+  +:+       +#+        */
+/*   By: mescobar <mescobar42@student.42perpigna    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/18 20:35:31 by qbanet            #+#    #+#             */
-/*   Updated: 2024/01/03 10:57:59 by mescobar         ###   ########.fr       */
+/*   Updated: 2024/01/05 18:33:23 by mescobar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -57,8 +57,8 @@ typedef struct s_camera
 	int				bpp;
 	int				line;
 	int				endian;
-	float			px;
-	float			py;
+	double			px;
+	double			py;
 	int				view;
 }					t_camera;
